+++ Plugin.cs
@@ -28,7 +28,7 @@ namespace HeavyItemSCPs
     {
         const string PLUGIN_GUID = "ProjectSCP.HeavyItemSCPs";
         const string PLUGIN_NAME = "HeavyItemSCPs";
-        const string PLUGIN_VERSION = "1.3.1";
+        const string PLUGIN_VERSION = "1.3.2";
 
         public static Plugin PluginInstance;
         public static ManualLogSource LoggerInstance;
@@ -47,25 +47,21 @@ namespace HeavyItemSCPs
         public static ConfigEntry<string> config427CustomLevelRarities;
         public static ConfigEntry<int> config427SCPDungeonRarity;
 
-        public static ConfigEntry<bool> configEnableOpenNecklace;
         public static ConfigEntry<float> configTimeToTransform;
-        public static ConfigEntry<float> configTransformOpenMultiplier;
-        public static ConfigEntry<int> configHealthPerSecondHolding;
         public static ConfigEntry<int> configHealthPerSecondOpen;
         public static ConfigEntry<float> configHoarderBugTransformTime;
         public static ConfigEntry<float> configBaboonHawkTransformTime;
         public static ConfigEntry<float> configOtherEnemyTransformTime;
         public static ConfigEntry<float> configTimeToSpawnSCP4271;
         public static ConfigEntry<int> configMaxSpawns;
-        public static ConfigEntry<bool> configIncreaseTimeWhenInInventory;
         public static ConfigEntry<bool> config427_500Compatibility;
-        public static ConfigEntry<bool> config427ResetTransformTimeAtOrbit;
 
         // SCP-427-1 Enemy Configs
         public static ConfigEntry<string> config4271LevelRarities;
         public static ConfigEntry<string> config4271CustomLevelRarities;
         public static ConfigEntry<int> config4271SCPDungeonRarity;
         public static ConfigEntry<DropMethod> config4271DropMethod;
+        public static ConfigEntry<int> config4271MaxHealth;
 
         // SCP-178 Configs
         public static ConfigEntry<bool> configEnableSCP178;
@@ -160,30 +156,26 @@ namespace HeavyItemSCPs
             // SCP-427
             configEnableSCP427 = Config.Bind("SCP-427", "Enable SCP-427", true, "Whether or not SCP-427 can spawn as scrap.");
             config427MinValue = Config.Bind("SCP-427", "Minimum value", 100, "The minimum value of SCP-427.");
-            config427MaxValue = Config.Bind("SCP-427", "Maximum value", 150, "The maximum value of SCP-427.");
+            config427MaxValue = Config.Bind("SCP-427", "Maximum value", 300, "The maximum value of SCP-427.");
             config427LevelRarities = Config.Bind("SCP-427 Rarities", "Level Rarities", "ExperimentationLevel:5, AssuranceLevel:7, VowLevel:10, OffenseLevel:15, AdamanceLevel:25, MarchLevel:15, RendLevel:30, DineLevel:35, TitanLevel:45, ArtificeLevel:25, EmbrionLevel:50, Modded:15", "Rarities for each level. See default for formatting.");
             config427CustomLevelRarities = Config.Bind("SCP-427 Rarities", "Custom Level Rarities", "Secret LabsLevel:50", "Rarities for modded levels. Same formatting as level rarities.");
             config427SCPDungeonRarity = Config.Bind("SCP-427 Rarities", "SCP Dungeon Rarity", 100, "The rarity of SCP-427 in the SCP Dungeon. Set to -1 to use level rarities.");
 
-            configEnableOpenNecklace = Config.Bind("SCP-427", "Enable opening necklace", true, "Whether or not players can open the necklace.");
-            configTimeToTransform = Config.Bind("SCP-427", "Time to transform", 120f, "How long a player can hold the necklace before they transform into SCP-427-1. Set to -1 to disable transforming.");
-            configTransformOpenMultiplier = Config.Bind("SCP-427", "Transform open multiplier", 3f, "The multiplier applied to the time it takes to transform when the necklace is opened. configTimeToTransform will be multiplied by this value.");
-            configHealthPerSecondHolding = Config.Bind("SCP-427", "Health per second holding", 2, "The health gained per second while holding SCP-427.");
-            configHealthPerSecondOpen = Config.Bind("SCP-427", "Health per second open", 7, "The health gained per second while opening SCP-427.");
+            configTimeToTransform = Config.Bind("SCP-427", "Time to transform", 30f, "How long a player can hold the necklace before they transform into SCP-427-1. Set to -1 to disable transforming.");
+            configHealthPerSecondOpen = Config.Bind("SCP-427", "Health per second open", 5, "The health gained per second while opening SCP-427.");
             configHoarderBugTransformTime = Config.Bind("SCP-427", "Hoarder bug transform time", 5f, "The time it takes for the hoarder bug to transform into SCP-427-1. Set to -1 to disable transforming.");
-            configBaboonHawkTransformTime = Config.Bind("SCP-427", "Baboon hawk transform time", 10f, "The time it takes for the baboon hawk to transform into SCP-427-1. Set to -1 to disable transforming.");
+            configBaboonHawkTransformTime = Config.Bind("SCP-427", "Baboon hawk transform time", 5f, "The time it takes for the baboon hawk to transform into SCP-427-1. Set to -1 to disable transforming.");
             configOtherEnemyTransformTime = Config.Bind("SCP-427", "Other enemy transform time", 10f, "The time it takes for the other enemies to transform into SCP-427-1. Set to -1 to disable transforming.");
             configTimeToSpawnSCP4271 = Config.Bind("SCP-427", "Time to spawn SCP-427-1", 300f, "The time it takes for SCP-427 to spawn SCP-427-1 when on the ground. Set to -1 to disable spawning from necklace.");
             configMaxSpawns = Config.Bind("SCP-427", "Max spawns", 1, "The maximum number of SCP-427-1 instances that can be spawned when SCP-427 is on the ground.");
-            configIncreaseTimeWhenInInventory = Config.Bind("SCP-427", "Increase time when in inventory", true, "Whether or not timeHoldingSCP427 should increase when SCP-427 is on hotbar but not being held.");
             config427_500Compatibility = Config.Bind("SCP-427", "SCP-500 compatibility", true, "Whether or not SCP-427 should be compatible with the SCP-500 mod. This will only work if you have the SCP-500 mod installed. If enabled, it will temporarily halt the transformation timer when holding or using SCP-427 when you take SCP-500.");
-            config427ResetTransformTimeAtOrbit = Config.Bind("SCP-427", "Reset transform time at orbit", false, "Whether or not SCP-427 should reset its transformation timer when orbiting.");
-
+            
             // SCP-427-1
             config4271LevelRarities = Config.Bind("SCP-427-1 Rarities", "Level Rarities", "ExperimentationLevel:0, AssuranceLevel:0, VowLevel:0, OffenseLevel:0, AdamanceLevel:0, MarchLevel:0, RendLevel:0, DineLevel:0, TitanLevel:0, ArtificeLevel:0, EmbrionLevel:0, Modded:0", "Rarities for each level. See default for formatting.");
             config4271CustomLevelRarities = Config.Bind("SCP-427-1 Rarities", "Custom Level Rarities", "Secret LabsLevel:0", "Rarities for modded levels. Same formatting as level rarities.");
             config4271SCPDungeonRarity = Config.Bind("SCP-427-1 Rarities", "SCP Dungeon Rarity", -1, "The rarity of SCP-427-1 in the SCP Dungeon. Set to -1 to use level rarities.");
-            config4271DropMethod = Config.Bind("SCP-427-1", "Drop method", DropMethod.DropNothing, "When the player is grabbed by SCP-427-1, they should drop: 0 = Drop nothing, 1 = Drop held item, 2 = Drop two-handed item, 3 = Drop all items.");
+            config4271DropMethod = Config.Bind("SCP-427-1", "Drop method", DropMethod.DropAllItems, "When the player is grabbed by SCP-427-1, they should drop: 0 = Drop nothing, 1 = Drop held item, 2 = Drop two-handed item, 3 = Drop all items.");
+            config4271MaxHealth = Config.Bind("SCP-427-1", "Max health", 50, "The maximum health of SCP-427-1.");
 
             // SCP-178
             configEnableSCP178 = Config.Bind("SCP-178", "Enable SCP-178", true, "Whether or not SCP-178 can spawn as scrap.");